as requested added authors to all added files. moved windowers to datautil
