+++ braindecode/datasets/dataset.py
@@ -2,6 +2,13 @@
 Dataset classes that handle data
 """
 
+# Authors: Hubert Banville <hubert.jbanville@gmail.com>
+#          Lukas Gemein <l.gemein@gmail.com>
+#          Simon Brandt <simonbrandt@protonmail.com>
+#          David Sabbagh <dav.sabbagh@gmail.com>
+#
+# License: BSD (3-clause)
+
 import numpy as np
 from torch.utils.data import Dataset
